Refactored Profile Page
