+++ Client/package.json
@@ -12,7 +12,8 @@
     "@emotion/cache": "^11.13.1",
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
-    "@mui/material": "^6.0.1",
+    "@mui/icons-material": "^6.0.2",
+    "@mui/material": "^6.0.2",
     "@mui/material-nextjs": "^6.0.1",
     "next": "14.2.5",
     "react": "^18",
