--- Client/src/app/profile/page.module.css
@@ -1,3 +1,0 @@
-.page {
-	margin: 5rem;
-}
